Make the autoreplace class callable, and update tests
